+++ marchenko3D/getFileInfo3D.c
@@ -10,7 +10,7 @@
 
 #define MAX(x,y) ((x) > (y) ? (x) : (y))
 #define MIN(x,y) ((x) < (y) ? (x) : (y))
-#define NINT(x) ((int)((x)>0.0?(x)+0.5:(x)-0.5))
+#define NINT(x) ((long)((x)>0.0?(x)+0.5:(x)-0.5))
 
 /**
 * gets sizes, sampling and min/max values of a SU file
@@ -24,13 +24,13 @@ void vmess(char *fmt, ...);
 void verr(char *fmt, ...);
 int optncr(int n);
 
-int getFileInfo3D(char *filename, int *n1, int *n2, int *n3, int *ngath, float *d1, float *d2, float *d3, float *f1, float *f2, float *f3, float *sclsxgxsygy, int *nxm)
+long getFileInfo3D(char *filename, long *n1, long *n2, long *n3, long *ngath, float *d1, float *d2, float *d3, float *f1, float *f2, float *f3, float *sclsxgxsygy, long *nxm)
 {
     FILE    *fp;
     size_t  nread, data_sz;
 	off_t   bytes, ret, trace_sz, ntraces;
-    int     sx_shot, sy_shot, gx_start, gx_end, gy_start, gy_end, itrace, one_shot, igath, end_of_file, fldr_shot;
-    int     verbose=1, igy, nsx, nsy;
+    long     sx_shot, sy_shot, gx_start, gx_end, gy_start, gy_end, itrace, one_shot, igath, end_of_file, fldr_shot;
+    long     verbose=1, igy, nsx, nsy;
     float   scl, *trace, dxsrc, dxrcv, dysrc, dyrcv;
     segy    hdr;
     
@@ -62,7 +62,7 @@ int getFileInfo3D(char *filename, int *n1, int *n2, int *n3, int *ngath, float *
 
     data_sz = sizeof(float)*(*n1);
     trace_sz = sizeof(float)*(*n1)+TRCBYTES;
-    ntraces  = (int) (bytes/trace_sz);
+    ntraces  = (long) (bytes/trace_sz);
 
     if (hdr.scalco < 0) scl = 1.0/fabs(hdr.scalco);
     else if (hdr.scalco == 0) scl = 1.0;
@@ -122,7 +122,7 @@ int getFileInfo3D(char *filename, int *n1, int *n2, int *n3, int *ngath, float *
 
 /* expensive way to find out how many gathers there are */
 
-//	fprintf(stderr, "ngath = %d dxrcv=%f d2=%f scl=%f \n", *ngath, dxrcv, *d2, scl);
+//	fprintf(stderr, "ngath = %li dxrcv=%f d2=%f scl=%f \n", *ngath, dxrcv, *d2, scl);
     if (*ngath == 0) {
 		*n2 = 0;
         *n3 = 0;
@@ -175,7 +175,7 @@ int getFileInfo3D(char *filename, int *n1, int *n2, int *n3, int *ngath, float *
                 dysrc = (float)(hdr.sy - sy_shot)*scl;
             }
             if (verbose>1) {
-                fprintf(stderr," . Scanning shot %d (%d) with %d traces dxrcv=%.2f dxsrc=%.2f %d %d dyrcv=%.2f dysrc=%.2f %d %d\n",sx_shot,igath,itrace,dxrcv*scl,dxsrc,gx_end,gx_start,dyrcv*scl,dysrc,gy_end,gy_start);
+                fprintf(stderr," . Scanning shot %li (%li) with %li traces dxrcv=%.2f dxsrc=%.2f %li %li dyrcv=%.2f dysrc=%.2f %li %li\n",sx_shot,igath,itrace,dxrcv*scl,dxsrc,gx_end,gx_start,dyrcv*scl,dysrc,gy_end,gy_start);
             }
             if (itrace != 0) { /* end of shot record */
                 fseeko( fp, -TRCBYTES, SEEK_CUR );
@@ -197,7 +197,7 @@ int getFileInfo3D(char *filename, int *n1, int *n2, int *n3, int *ngath, float *
 		*ngath = ntraces/((*n2)*(*n3));
     }
 //    *nxm = NINT((*xmax-*xmin)/dxrcv)+1;
-	*nxm = (int)ntraces;
+	*nxm = (long)ntraces;
 
     fclose( fp );
     free(trace);
@@ -205,13 +205,13 @@ int getFileInfo3D(char *filename, int *n1, int *n2, int *n3, int *ngath, float *
     return 0;
 }
 
-int disp_fileinfo3D(char *file, int n1, int n2, int n3, float f1, float f2, float f3, float d1, float d2, float d3, segy *hdrs)
+long disp_fileinfo3D(char *file, long n1, long n2, long n3, float f1, float f2, float f3, float d1, float d2, float d3, segy *hdrs)
 {
 	vmess("file %s contains", file);
-    vmess("*** n1 = %d n2 = %d n3 = %d ntftt=%d", n1, n2, n3, optncr(n1));
+    vmess("*** n1 = %li n2 = %li n3 = %li ntftt=%li", n1, n2, n3, (long)optncr((int)n1));
 	vmess("*** d1 = %.5f d2 = %.5f d3 = %.5f", d1, d2, d3);
 	vmess("*** f1 = %.5f f2 = %.5f f3 = %.5f", f1, f2, f3);
-	vmess("*** fldr = %d sx = %d sy = %d", hdrs[0].fldr, hdrs[0].sx, hdrs[0].sy);
+	vmess("*** fldr = %li sx = %li sy = %li", hdrs[0].fldr, hdrs[0].sx, hdrs[0].sy);
 
 	return 0;
 }
